Organize "action" crate
